+++ P1/src/selection_methods/estocastico.java
@@ -2,7 +2,7 @@ package selection_methods;
 
 public class estocastico implements selection_method{
     public tabla_frecuencias t;
-    public int[] chooseEntities(int[] fitness){
+    public int[] chooseEntities(double[] fitness){
         int[] selected = new int[fitness.length];
         t = new tabla_frecuencias(fitness);
         float separacion = 1.0f / fitness.length;
